+++ scripts/headless_test.js
@@ -5451,6 +5451,7 @@ function check(name, ok, extra) {
   await require("./engine_sound_checks")({ newPage, check, shots: SHOTS });
   await require("./lights_police_checks")({ newPage, check, shots: SHOTS, viewports: [[1024,768],[768,1024]] });
   await require("./car_feel_checks")({ newPage, check, shots: SHOTS, viewports: [[1024,768],[390,844]] });
+  await require("./cleanup_checks")({ newPage, check, shots: SHOTS, viewports: [[1024,768],[820,1180]] });
 
   await browser.close();
   server.kill();